move main.ts into src/code/electron-main
